agent/pyproject.toml
[project]
name = "relocation-quest-agent"
version = "0.1.0"
description = "ATLAS - Relocation Quest voice assistant powered by Pydantic AI"
requires-python = ">=3.11"
dependencies = [
    "pydantic-ai[google]>=0.0.40",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
    "asyncpg>=0.29.0",
    "google-generativeai>=0.8.6",
    "zep-cloud>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
